+++ src/autogluon/cloud/backend/constant.py
@@ -1,2 +1,3 @@
 SAGEMAKER = "sagemaker"
+TIMESERIES_SAGEMAKER = "timeseries_sagemaker"
 RAY = "ray"
